moved TrackableObject to main.py
