--- gen2-cumulative-object-counting/trackable_object.py
@@ -1,11 +1,0 @@
-# from https://www.pyimagesearch.com/2018/08/13/opencv-people-counter/
-class TrackableObject:
-    def __init__(self, objectID, centroid):
-        # store the object ID, then initialize a list of centroids
-        # using the current centroid
-        self.objectID = objectID
-        self.centroids = [centroid]
-
-        # initialize a boolean used to indicate if the object has
-        # already been counted or not
-        self.counted = False
